+++ hpc-engine/cpca_kernels.cu
@@ -225,6 +225,7 @@ extern "C" int compute_derivative_real_vector_f(double *fx, double *fy, double *
 extern "C" int compute_laplace_real_f(double *f, double *laplace_f, int nthreads);
 extern "C" int compute_divergence_real_vector_f(double *fx, double *fy, double *fz, double *divf, int nthreads);
 extern "C" int high_frequency_filter_d(double *in, double *out, double fd_mu, double fd_T, int nthreads);
+extern "C" int high_frequency_filter_c(Complex *in, Complex *out, double fd_mu, double fd_T, int nthreads);
 extern "C" int high_frequency_filter_massive_d(int n, double *in, double *out, double fd_mu, double fd_T, int nthreads);
 
 __global__ void kernel_apply_hamiltonian(int it, wslda_potential h_potentials,
@@ -694,8 +695,7 @@ extern "C" int apply_hamiltonian(int it, int n, cufftDoubleComplex *wf_in, cufft
     }
 
     // filtering of mean-fields
-    double Emax =  0.5*M_PI*M_PI/DX/DX; // E_max = p_max^2 / 2m, where: p_max is maximum momentum on the lattice, p_max=M_PI/dx
-    if(md.hkf_mu<9.9*Emax)
+    if(md.hkf_mode>=1)
     {
         ierr = high_frequency_filter_massive_d(2, V_a, V_a, md.hkf_mu, md.hkf_T, nthreads);
         if(ierr!=0) return ierr;
@@ -722,6 +722,12 @@ extern "C" int apply_hamiltonian(int it, int n, cufftDoubleComplex *wf_in, cufft
     if(ierr!=0) return ierr+400;
 #endif
 
+    if(md.hkf_mode>=2)
+    {
+        ierr = high_frequency_filter_c(potsall.delta, potsall.delta, md.hkf_mu, md.hkf_T, nthreads);
+        if(ierr!=0) return ierr;
+    }
+
 #ifdef BDG_MODE
     // Step 3: apply hamiltonian
     kernel_apply_hamiltonian_bdg<<<nblocks, nthreads>>>(it,
@@ -742,7 +748,7 @@ extern "C" int apply_hamiltonian(int it, int n, cufftDoubleComplex *wf_in, cufft
 
 #ifdef CURRENT_CORRECTIONS
     // filtering of vector potentials
-    if(md.hkf_mu<9.9*Emax)
+    if(md.hkf_mode>=1)
     {
         ierr = high_frequency_filter_massive_d(6, potsall.A_a_x, potsall.A_a_x, md.hkf_mu, md.hkf_T, nthreads);
         if(ierr!=0) return ierr;